Split current NVM code into three separate modules
