+++ src/CMakeLists.txt
@@ -237,6 +237,12 @@ add_library(libavrdude
     updi_link.h
     updi_nvm.c
     updi_nvm.h
+    updi_nvm_v0.c
+    updi_nvm_v0.h
+    updi_nvm_v2.c
+    updi_nvm_v2.h
+    updi_nvm_v3.c
+    updi_nvm_v3.h
     updi_readwrite.c
     updi_readwrite.h
     updi_state.c
