+++ src/updi_constants.h
@@ -106,51 +106,4 @@
 
 #define UPDI_RESET_REQ_VALUE  0x59
 
-// FLASH CONTROLLER
-#define UPDI_NVMCTRL_CTRLA    0x00
-#define UPDI_NVMCTRL_CTRLB    0x01
-#define UPDI_NVMCTRL_STATUS   0x02
-#define UPDI_NVMCTRL_INTCTRL  0x03
-#define UPDI_NVMCTRL_INTFLAGS 0x04
-#define UPDI_NVMCTRL_DATAL    0x06
-#define UPDI_NVMCTRL_DATAH    0x07
-#define UPDI_NVMCTRL_ADDRL    0x08
-#define UPDI_NVMCTRL_ADDRH    0x09
-
-// NVMCTRL v0 CTRLA
-#define UPDI_V0_NVMCTRL_CTRLA_NOP              0x00
-#define UPDI_V0_NVMCTRL_CTRLA_WRITE_PAGE       0x01
-#define UPDI_V0_NVMCTRL_CTRLA_ERASE_PAGE       0x02
-#define UPDI_V0_NVMCTRL_CTRLA_ERASE_WRITE_PAGE 0x03
-#define UPDI_V0_NVMCTRL_CTRLA_PAGE_BUFFER_CLR  0x04
-#define UPDI_V0_NVMCTRL_CTRLA_CHIP_ERASE       0x05
-#define UPDI_V0_NVMCTRL_CTRLA_ERASE_EEPROM     0x06
-#define UPDI_V0_NVMCTRL_CTRLA_WRITE_FUSE       0x07
-
-// NVMCTRL v2 CTRLA
-#define UPDI_V2_NVMCTRL_CTRLA_NOCMD              0x00
-#define UPDI_V2_NVMCTRL_CTRLA_FLASH_WRITE        0x02
-#define UPDI_V2_NVMCTRL_CTRLA_FLASH_PAGE_ERASE   0x08
-#define UPDI_V2_NVMCTRL_CTRLA_EEPROM_ERASE_WRITE 0x13
-#define UPDI_V2_NVMCTRL_CTRLA_CHIP_ERASE         0x20
-#define UPDI_V2_NVMCTRL_CTRLA_EEPROM_ERASE       0x30
-
-// NVMCTRL v3 CTRLA
-#define UPDI_V3_NVMCTRL_CTRLA_NOCMD                    0x00
-#define UPDI_V3_NVMCTRL_CTRLA_NOP                      0x01
-#define UPDI_V3_NVMCTRL_CTRLA_FLASH_PAGE_WRITE         0x04
-#define UPDI_V3_NVMCTRL_CTRLA_FLASH_PAGE_ERASE_WRITE   0x05
-#define UPDI_V3_NVMCTRL_CTRLA_FLASH_PAGE_ERASE         0x08
-#define UPDI_V3_NVMCTRL_CTRLA_FLASH_PAGE_BUFFER_CLEAR  0x0F
-#define UPDI_V3_NVMCTRL_CTRLA_EEPROM_PAGE_WRITE        0x14
-#define UPDI_V3_NVMCTRL_CTRLA_EEPROM_PAGE_ERASE_WRITE  0x15
-#define UPDI_V3_NVMCTRL_CTRLA_EEPROM_PAGE_ERASE        0x17
-#define UPDI_V3_NVMCTRL_CTRLA_EEPROM_PAGE_BUFFER_CLEAR 0x1F
-#define UPDI_V3_NVMCTRL_CTRLA_CHIP_ERASE               0x20
-#define UPDI_V3_NVMCTRL_CTRLA_EEPROM_ERASE             0x30
-
-#define UPDI_NVM_STATUS_WRITE_ERROR 2
-#define UPDI_NVM_STATUS_EEPROM_BUSY 1
-#define UPDI_NVM_STATUS_FLASH_BUSY  0
-
 #endif /* updi_constants_h */
